Decouple content allocation and realization.
This makes it possible to allocate the content storage identity
(necessary for creating the URI) before the content itself is realized.
Meaning actions are now fully asynchronous.

The trade-off here is that errors that occur while realizing the content
are then only visible by attempting to retrieve the content, instead of
before the action has been performed.